refactor(types): separate type checking logic into check-base.ts
- Move core type checking logic to check-base.ts
- Simplify check.ts to be a thin wrapper
- Pass type definitions as arguments for better dependency management
