+++ packages/@markuplint/types/src/index.ts
@@ -1,3 +1,4 @@
 export * from './whatwg/is-custom-element-name.js';
 export * from './check.js';
+export * from './check-base.js';
 export * from './types.js';
