+++ packages/@markuplint/types/src/list.ts
@@ -1,12 +1,12 @@
-import type { Result } from './types.js';
+import type { Defs, Result } from './types.js';
 import type { List } from './types.schema.js';
 import type { ReadonlyDeep } from 'type-fest';
 
-import { check } from './check.js';
+import { checkBase } from './check-base.js';
 import { matched } from './match-result.js';
 import { Token, TokenCollection } from './token/index.js';
 
-export function checkList(value: string, type: ReadonlyDeep<List>, ref?: string, cache = true): Result {
+export function checkList(value: string, type: ReadonlyDeep<List>, defs: Defs, ref?: string, cache = true): Result {
 	const tokens = new TokenCollection(value, type);
 
 	const matches = tokens.check({ ref });
@@ -18,7 +18,7 @@ export function checkList(value: string, type: ReadonlyDeep<List>, ref?: string,
 	const identTokens = tokens.getIdentTokens();
 
 	for (const token of identTokens) {
-		const res = check(token.value, type.token, ref, cache);
+		const res = checkBase(token.value, type.token, defs, ref, cache);
 		if (!res.matched) {
 			const { offset, line, column } = Token.shiftLocation(token, res.offset);
 			return {
